requirements.txt
Django==1.9.7
django-filebrowser==3.7.1
django-grappelli==2.8.1
Pillow==2.2.1